refactor(9-2b): promote 9-2a + 8-1b helpers to pub(crate) util
T1 (zip dep) + T5(util) refactor préparatoire avant les serializers
CSV et le handler `routes/exports.rs` de Story 9-2b :

- `crates/kesh-api/src/util.rs` (nouveau) — slugify, build_content_disposition,
  percent_encode_filename, ascii_fallback_filename, hex_encode + nouveau
  helper map_language_to_bcp47 (extrait du match inline `load_pdf_context`
  ground-truth `reports.rs:573` per Pass 3 ECH3-H2).
- `routes/reports.rs` — remplace helpers privés par `crate::util::*`,
  tests slugify/content-disposition/percent migrés vers `util::tests`.
- `routes/bank_imports.rs` — supprime `hex_encode` local, call `util::hex_encode`.
- `Cargo.toml` — ajoute `zip = "2"` (default-features=false + deflate),
  seule nouvelle dep (sha2 + csv déjà présents 9-2a/8-1b).
- sprint-status.yaml + story file — status `ready-for-dev → in-progress`.

Tests : 18 ok (5 build_filename/validate_format reports + 13 util — slugify×4,
content-disposition×2, percent-encode×2, map_language_to_bcp47×2, hex_encode×3).
20 `config::tests::*` régressions pré-existantes héritées 9-2a inchangées.
0 régression bank_imports.
