+++ _bmad-output/implementation-artifacts/9-2b-export-global-zip.md
@@ -1,6 +1,6 @@
 # Story 9.2b: Export global ZIP (souveraineté des données)
 
-Status: ready-for-dev
+Status: in-progress
 
 <!-- Note: Validation is optional. Run validate-create-story for quality check before dev-story. -->
